+++ interface/app/$libraryId/Explorer/TopBarOptions.tsx
@@ -1,6 +1,5 @@
 import {
 	ArrowClockwise,
-	FolderPlus,
 	Icon,
 	Key,
 	MonitorPlay,
@@ -36,29 +35,7 @@ export const useExplorerTopBarOptions = () => {
 	const explorer = useExplorerContext();
 	const controlIcon = useKeyMatcher('Meta').icon;
 	const settings = explorer.useSettingsSnapshot();
-
 	const rescan = useQuickRescan();
-
-	const createFolder = useLibraryMutation(['files.createFolder'], {
-		onError: (e) => {
-			toast.error({ title: 'Error creating folder', body: `Error: ${e}.` });
-			console.error(e);
-		},
-		onSuccess: (folder) => {
-			toast.success({ title: `Created new folder "${folder}"` });
-			rescan();
-		}
-	});
-	const createEphemeralFolder = useLibraryMutation(['ephemeralFiles.createFolder'], {
-		onError: (e) => {
-			toast.error({ title: 'Error creating folder', body: `Error: ${e}.` });
-			console.error(e);
-		},
-		onSuccess: (folder) => {
-			toast.success({ title: `Created new folder "${folder}"` });
-			rescan();
-		}
-	});
 
 	const viewOptions = useMemo(
 		() =>
@@ -134,26 +111,6 @@ export const useExplorerTopBarOptions = () => {
 
 	const toolOptions = [
 		(parent?.type === 'Location' || parent?.type === 'Ephemeral') && {
-			toolTipLabel: 'New Folder',
-			icon: <FolderPlus className={TOP_BAR_ICON_STYLE} />,
-			onClick: () => {
-				if (parent?.type === 'Location') {
-					createFolder.mutate({
-						location_id: parent.location.id,
-						sub_path: path || null,
-						name: null
-					});
-				} else {
-					createEphemeralFolder.mutate({
-						path: parent?.path,
-						name: null
-					});
-				}
-			},
-			individual: true,
-			showAtResolution: 'xs:flex'
-		},
-		{
 			toolTipLabel: 'Key Manager',
 			icon: <Key className={TOP_BAR_ICON_STYLE} />,
 			popOverComponent: <KeyManager />,
